refactor(mdx): create transferMarkdownImageNodeToJSX
